ci: migrate to ci v1
