+++ pyproject.toml
@@ -1,6 +1,42 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
+# Dependencies
+[project]
+name = "kiali-k8s"
+version = "0.0"
+requires-python = "~=3.8"
+
+dependencies = [
+    "cosl",
+    "ops",
+    "pydantic",
+    "requests",
+    "observability-charm-tools@git+https://github.com/canonical/observability-charm-tools@main",
+]
+
+[project.optional-dependencies]
+dev = [
+    # Liniting
+    "ruff",
+    "black",
+    "codespell",
+    # Static
+    "pyright",
+    # Unit
+    "pytest",
+    "pytest-mock",
+    "coverage[toml]",
+    "ops-scenario",
+    # Integration
+    "juju",
+    "lightkube",
+    "minio",
+    "pytest-operator",
+    "requests",
+    "pytest-asyncio==0.21.2",
+]
+
 # Testing tools configuration
 [tool.coverage.run]
 branch = true
