+++ tox.ini
@@ -3,90 +3,68 @@
 
 [tox]
 skip_missing_interpreters = True
-envlist = lint, static-{charm}, unit, scenario
+envlist = lint, static-{charm}, unit
 
 [vars]
 src_path = {toxinidir}/src
 tst_path = {toxinidir}/tests
 all_path = {[vars]src_path} {[vars]tst_path}
+uv_flags = --frozen --isolated --extra=dev
 
 [testenv]
+allowlist_externals = uv
 basepython = python3
 setenv =
   PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
   PYTHONBREAKPOINT=ipdb.set_trace
   PY_COLORS=1
+passenv =
+  PYTHONPATH
+  CHARM_PATH
+
+[testenv:lock]
+deps = uv
+description = Update uv.lock with the latest deps
+commands =
+  uv lock --upgrade --no-cache
 
 [testenv:fmt]
-description = Apply coding style standards to code
-deps =
-    black
-    ruff
+description = Format the code
 commands =
-    ruff check --fix {[vars]all_path}
-    black {[vars]all_path}
+    uv run {[vars]uv_flags} ruff check --fix-only {[vars]all_path}
 
 [testenv:lint]
-description = Check code against coding style standards
-deps =
-    black
-    ruff
-    codespell
+description = Lint the code
 commands =
-    codespell {toxinidir} --skip {toxinidir}/src/manifests
-    ruff check {[vars]all_path}
-    black --check --diff {[vars]all_path}
+    uv run {[vars]uv_flags} ruff check {[vars]all_path}
 
-[testenv:static-{charm,lib}]
-skip_install=True
-description = Run static analysis checks
-deps =
-    pyright
-    -r {toxinidir}/requirements.txt
+[testenv:static]
+description = Run static checks
+allowlist_externals =
+  {[testenv]allowlist_externals}
+  /usr/bin/env
 commands =
-    charm: pyright {[vars]src_path}
-    lib:
+    uv run {[vars]uv_flags} pyright {[vars]all_path}
+    /usr/bin/env sh -c 'for m in $(git diff main --name-only {[vars]lib_path}); do if ! git diff main $m | grep -q "+LIBPATCH\|+LIBAPI"; then echo "You forgot to bump the version on $m!"; exit 1; fi; done'
 
 [testenv:unit]
 description = Run unit tests
-deps =
-    pytest
-    pytest-mock
-    coverage[toml]
-    -r {toxinidir}/requirements.txt
+setenv =
+  {[testenv]setenv}
+  JUJU_VERSION=3.0.3
+passenv = 
+    PYTHONPATH
+allowlist_externals =
+  {[testenv]allowlist_externals}
+  /usr/bin/env
 commands =
-    coverage run \
-      --source={[vars]src_path} \
-      -m pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/unit
-    coverage report -m
-
-[testenv:scenario]
-description = Run scenario tests
-deps =
-    -r{toxinidir}/requirements.txt
-    pytest
-    ops-scenario
-commands =
-    pytest -vv --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/scenario
+    uv run {[vars]uv_flags} coverage run --source={[vars]src_path} -m pytest \
+        {[vars]tst_path}/unit {posargs}
+    uv run {[vars]uv_flags} coverage report
 
 [testenv:integration]
 description = Run integration tests
 passenv =
     KIALI_CHARM
-deps =
-    pytest
-    pytest-asyncio==0.21.2
-    juju
-    lightkube
-    minio
-    pytest-operator
-    requests
-    -r {toxinidir}/requirements.txt
 commands =
-    pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tst_path}/integration
-
-[testenv:check]
-skip_install=True
-depends =
-    lint
-    static
+    uv run {[vars]uv_flags} pytest --exitfirst {[vars]tst_path}/integration {posargs}
